requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
